Refactor Nearby Wi-Fi Devices permission logic
